requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.2
frozenlist==1.8.0
future==1.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
line-bot-sdk==3.19.2
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
requests==2.32.5
schedule==1.2.2
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.22.0
